🚚 Move text escapers into internals of /arbitrary
