+++ src/arbitrary/_internals/helpers/TextEscaper.ts
